requirements.txt
Hypercorn==0.11.2
PyYAML==5.4.1
fastapi==0.68.1
tortoise-orm==0.17.7
